dded webhelper and a new feature test
Added two new webhelpers to sign up and add an example post. Created a new feature test to cehck the username is displayed when a post is made. Test currently failing. Code added to the post controller to get the username information when a post is made. Todo: need to update the index view to show the username now that we understand how to egt it from the post.